requirements.txt
# pip install -r requirements.txt
langgraph
python-dotenv
langchain-openai
langchain
langchain-core
astrapy